renamed all solver api so not to clash with their python package names
